Tests/LibWeb/Text/expected/abspos-flex-calc-insets-definite-height.txt

container height: 120
item height: 60
item height is 50% of resolved inset height: true
